requirements.txt
cffi==2.0.0
flatbuffers==25.12.19
mpmath==1.3.0
numpy==2.4.2
onnxruntime==1.24.1
packaging==26.0
piper-tts==1.4.1
protobuf==6.33.5
pycparser==3.0
sounddevice==0.5.5
sympy==1.14.0
